* Reorder errors fixed

+++ Source/Gorgon/Animation/Animation.cpp
@@ -125,7 +125,7 @@ namespace Gorgon :: Animation {
                     if(!length && animations.GetSize()) {
                         length = animations[0].GetDuration();
                         
-                        if(length == -1) {
+                        if(length == (unsigned)-1) {
                             length = 0;
                         }
                     }
@@ -174,7 +174,7 @@ namespace Gorgon :: Animation {
             if(!length && speed < 0 && animations.GetSize()) {
                 length = animations[0].GetDuration();
                 
-                if(length == -1) {
+                if(length == (unsigned)-1) {
                     length = 0;
                 }
             }
